add /allocate and /charge endpoints
The /allocate endpoint does the ticket allocation. Right now it's pretty
basic, so I will deal with some of the edge cases later on down the
line.

This also adds the `/charge` endpoint, which uses the ticket guid from
the cookie to look up the ticket. This is where I'll allow the user to
checkout with Stripe later.